% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Algorithm.Niche.Model.R
\docType{class}
\name{Algorithm.Niche.Model-class}
\alias{Algorithm.Niche.Model-class}
\title{An S4 class to represent a specie distribution model of one algorithm}
\description{
This is an S4 class to represent a specie distribution model of one algorithm
(among generalized linear model, general additive model, multivariate
adpatative splines, generalized boosted regression models, classification tree
analysis, random forest, maximum entropy, artificial neural network, and
support vector machines). It can be obtain with \code{\link{Modelling}}.
}
\section{Slots}{

\describe{
\item{\code{name}}{character. Name of the model (by default
Specie.Algorithm.Niche.Model)}

\item{\code{projection}}{raster. Habitat suitability map of the model}

\item{\code{evaluation}}{data frame. Evaluation of the model (threshold, AUC, omission
rate, sensitivity, specificity, correct proportion and Kappa)}

\item{\code{variables.importance}}{data frame. Relative percentage of importance for each variable used in the model}

\item{\code{data}}{data frame. Data used to realized the model}

\item{\code{parameters}}{data frame. Parameters used to realized the model}
}}
\seealso{
\linkS4class{Ensemble.Niche.Model} an S4 class for ensemble models,
 and \linkS4class{Stack.Species.Ensemble.Niche.Model} an S4 class for stack
 species enemble models.
}

