% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ensemble.R
\docType{methods}
\name{ensemble,Algorithm.Niche.Model-method}
\alias{ensemble,Algorithm.Niche.Model-method}
\title{Function to assemble different algorithms models in one ensemble model}
\usage{
\S4method{ensemble}{Algorithm.Niche.Model}(x, ..., name = NULL,
  ensemble.metric = c("AUC"), ensemble.thresh = c(0.75), weight = T,
  thresh = 1001, uncertainity = T)
}
\arguments{
\item{x,...}{Algorithm.Niche.Model. Algortihms models to assemble.}

\item{name}{character. Optionnal name given to the final Ensemble.Niche.Model producted.}

\item{thresh}{numeric. binary map threshold computing precision parmeter, the higher it is the more accurate is the threshold but the longer is the modelling evaluation step !}

\item{uncertainity}{logical. If false uncertainity mapping and algorithms correlation matrix is not computed.}

\item{AUCthresh}{numeric. Value under wich models will be rejected (corresponding to the metric parameter).}
}
\value{
an S4 Ensemble.Niche.Model Class object viewable with plot method
}
\description{
This is a function to assemble several algorithms models in one ensemble model. It takes in inputs several S4 Algorithm.Niche.Model class objects obtained with the Modelling function. It returns an S4 Ensemble.Niche.Model class object containing the habitat suitability map, the binary map, and the uncertainty map based on the habitat suitability map variance inter algorithms and all evaluation tables comming with (model evaluation, algorithms evaluation, algorithms correlation matrix and variables importance).
}
\examples{
\dontrun{
ensemble(GLM1, GLM2, GAM1, GAM2)
}
}
\seealso{
Ensemble.Modelling for stack ensemble modelling with multiple algorithms
}

