% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Ensemble.Niche.Model.R
\docType{class}
\name{Ensemble.Niche.Model-class}
\alias{Ensemble.Niche.Model-class}
\title{An S4 class to represent a specie distribution ensemble model of multiple
algorithms}
\description{
This is an S4 class to represent a specie distribution ensemble model of
multiple algorithms (among generalized linear model, general additive model,
multivariate adpatative splines, generalized boosted regression models,
classification tree analysis, random forest, maximum entropy, artificial
neural network, and support vector machines). It can be obtain with
\code{\link{Ensemble.Modelling}} or \code{\link{ensemble}}.
}
\section{Slots}{

\describe{
\item{\code{uncertainity}}{raster. Intermodel variance map}

\item{\code{algorithm.correlation}}{data.frame. Correlation matrix inter algorithms}

\item{\code{algorithm.evaluation}}{calculated with pearson coefficient on habitat suitability map}
}}
\seealso{
\linkS4class{Algorithm.Niche.Model} an S4 class for specie models
 with only one algorithm, and
 \linkS4class{Stack.Species.Ensemble.Niche.Model} an S4 class for stack
 species enemble models.
}

