% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ensemble.R
\docType{methods}
\name{ensemble}
\alias{ensemble}
\alias{ensemble,Algorithm.Niche.Model-method}
\alias{sum,Algorithm.Niche.Model-method}
\title{Method to assemble different algorithms models in one ensemble model}
\usage{
ensemble(x, ..., name = NULL, ensemble.metric = c("AUC"),
  ensemble.thresh = c(0.75), weight = T, thresh = 1001,
  uncertainity = T, verbose = T, GUI = F)

\S4method{ensemble}{Algorithm.Niche.Model}(x, ..., name = NULL,
  ensemble.metric = c("AUC"), ensemble.thresh = c(0.75), weight = T,
  thresh = 1001, uncertainity = T, verbose = T, GUI = F)

\S4method{sum}{Algorithm.Niche.Model}(x, ..., name = NULL,
  ensemble.metric = c("AUC"), ensemble.thresh = c(0.75), weight = T,
  thresh = 1001, format = T, verbose = T, na.rm = F)
}
\arguments{
\item{x,...}{Algorithm.Niche.Model. Algortihms models to assemble.}

\item{name}{character. Optionnal name given to the final Ensemble.Niche.Model
producted.}

\item{ensemble.metric}{character. Metric used to compute the selection among
algorithms different models (see details below)}

\item{ensemble.thresh}{numeric. Threshold associated with the metric used to
compute the selection.}

\item{weight}{logical. Choose if the model are weighted or not by the
selection associated metrics mean.}

\item{thresh}{numeric. Binary map threshold computing precision parmeter, the
higher it is the more accurate is the threshold but the longer is the
modelling evaluation step !}

\item{uncertainity}{logical. If false uncertainity mapping and algorithms
correlation matrix is not computed.}

\item{verbose}{logical. If true allow the function to print text in the
console}

\item{GUI}{logical. Don't take that argument into account (parameter for the
 user interface) !}

\item{format}{logical. Should always kept to true, it block the models sum to
only model from the same algorithm}

\item{na.rm}{logical. Should NA be removed in models sum}
}
\value{
an S4 \linkS4class{Ensemble.Niche.Model} Class object viewable with
 \code{\link{plot.model}} method
}
\description{
This is a method to assemble several algorithms models in one ensemble model.
It takes in inputs several S4 \linkS4class{Algorithm.Niche.Model} class
objects obtained with the \code{\link{Modelling}} function. It returns an S4
\linkS4class{Ensemble.Niche.Model} class object containing the habitat
suitability map, the binary map, and the uncertainty map based on the habitat
suitability map variance inter algorithms and all evaluation tables comming
with (model evaluation, algorithms evaluation, algorithms correlation matrix
and variables importance).
}
\details{
Ensemble metric (Metric used to compute the selection among
 algorithms different models) can be choosed among those : \describe{
 \item{AUC}{Area under the receiving operative curve (ROC)}
 \item{Kappa}{Kappa metric issued from the confusion matrix}
 \item{sensitivity}{Sensitivity metric issued from the confusion matrix}
 \item{specificity}{Specificity metric issued from the confusion matrix}
 \item{prop.correct}{Correct predicted occurences proportion issued from the
 confusion matrix} }

 Sum method is an intermediate method to only sum multiple algoritms from one
 origin in one model but without computing uncertainty mapping, algorithms
 correlation and evaluation and model evaluation.
}
\examples{
\dontrun{
ensemble(GLM1, GLM2, GAM1, GAM2)
}
}
\seealso{
\code{\link{Ensemble.Modelling}} for stack ensemble modelling with
 multiple algorithms
}

