% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/evaluate.R
\docType{methods}
\name{evaluate,Niche.Model-method}
\alias{evaluate,Niche.Model-method}
\title{Method to evaluate a model}
\usage{
\S4method{evaluate}{Niche.Model}(obj, thresh = 1001, metric = "SES")
}
\arguments{
\item{obj}{Algorithm.Niche.Model or Ensemble.Niche.Model. Model to evaluate}

\item{thresh}{numeric. Binary map threshold computing precision parmeter, the
 higher it is the more accurate is the threshold but the longer is the
 modelling evaluation step !}

\item{metric}{character. Method used to calculate the optimal threshold to
compute the confusion matrix used for the evaluation of the model (see
details below)}
}
\value{
S4 \linkS4class{Algorithm.Niche.Model} or
 \linkS4class{Ensemble.Niche.Model} classes objects. viewable with
 \code{\link{plot.model}} method depending on the input.
}
\description{
This is a method to evaluate models from S4
\linkS4class{Algorithm.Niche.Model} or \linkS4class{Ensemble.Niche.Model}
classes objects.
}
\details{
\strong{Metric :} choice of the method used to compute binary map
threshold and confusion matrix : \describe{ \item{Kappa}{maximizes the model
Kappa value} \item{TSS}{\strong{True Skill Statistic} maximizes the
sensitivity and specificity sum} \item{CCR}{maximizes the correct predicted
observations proportion} \item{SES}{using the sensitivty specificity
equality} \item{LW}{using the lowest occurence prediction probability}
\item{ROC}{minimizing the distance between the ROC plot (receiving operative
curve) and the upper left coin (1,1)} }
}
\examples{
\dontrun{
evaluate(GLM)
}
}
\seealso{
\code{\link{Modelling}} for specie distribution modelling with one
 algorithm, \code{\link{Ensemble.Modelling}} for specie ensemble modelling
 with multiple algorithms
}

