% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/load.model.R
\name{load.model}
\alias{load.enm}
\alias{load.model}
\alias{load.stack}
\title{Load models functions}
\usage{
load.enm(name, directory = getwd())

load.stack(name = "Stack", directory = getwd(), GUI = F)
}
\arguments{
\item{name}{character. Folder name of the saved model.}

\item{directory}{character. Path to the directory containing the saved model
folder, by default the current directory.}

\item{GUI}{logical. Don't take that argument into account (parameter for the user interface) !}
}
\value{
The corresponding model object
}
\description{
Allow to load S4 \linkS4class{Ensemble.Niche.Model} and
\linkS4class{Stack.Species.Ensemble.Niche.Model} objects saved with their
respective save functions
}
\seealso{
\code{\link{save.model}}
}

