% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/load.occ.R
\name{load.occ}
\alias{load.occ}
\title{Load occurences data}
\usage{
load.occ(directory = getwd(), Env, file = NULL, ..., Xcol = "Longitude",
  Ycol = "Latitude", Spcol = NULL, GeoRes = T,
  reso = max(res(Env@layers[[1]])), verbose = T, GUI = F)
}
\arguments{
\item{directory}{character. Directory where is the occurences table file}

\item{Env}{raster stack. Environment vraiables as a raster stack used to do
the spatial thinning, it can be the result of the \code{\link{load.var}}
function}

\item{file}{character. File containing the occurences table, if NULL (default)
.csv file present in the directory will be loaded}

\item{...}{additional parameters gave to \code{\link[utils]{read.csv}}}

\item{Xcol}{character. Name of the X coordinates column (Longitude)}

\item{Ycol}{character. Name of the Y coordinates column (Latitude)}

\item{Spcol}{character. Name of the column containing the species name or ID}

\item{GeoRes}{logical. Geographical thinning will be computed on occurences by
species to avoid geographical bias in models}

\item{reso}{numeric. Resolution used to do the geographical thinnig, by
default the environmental variables raster stack resolution (Env)}

\item{verbose}{logical. If true allow the function to print text in the
console}

\item{GUI}{logical. Don't take that argument into account (parameter for the user interface) !}
}
\value{
A data frame containing the occurences data set (spatially thinned or
 not)
}
\description{
Function to load occurences data from a table to do \code{\link{Modelling}},
\code{\link{Ensemble.Modelling}} or \code{\link{Stack.Modelling}}.
}
\examples{
\dontrun{
load.occ(directory)
}
}
\seealso{
\code{\link{load.var}} to load environmental variables
}

