% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/load.var.R
\name{load.var}
\alias{load.var}
\title{Load environmental variables data}
\usage{
load.var(directory = getwd(), files = NULL, format = c(".grd", ".tif",
  ".asc", ".sdat", ".rst", ".nc", ".tif", ".envi", ".bil", ".img"),
  factors = NULL, Norm = T, tmp = T, verbose = T, GUI = F)
}
\arguments{
\item{directory}{character. Directory where are the environmental data
variables files}

\item{files}{character. Files containing the environmental data variables, if
NULL (default) all files with the precised format present in the directory
will be loaded.}

\item{format}{character. Format used to load environmental data variables
files (among .grd, .tif, .asc, .sdat, .rst, .nc, .tif, .envi, .bil, .img)}

\item{factors}{character. Environmental data variables which should be
considered as factor variables}

\item{Norm}{logical. If true environmental data are normalized.}

\item{tmp}{logical. If true loaded environmental data variables rasters are
read in temporary file avoiding to overload the random access memory}

\item{verbose}{logical. If true allow the function to print text in the
console}

\item{GUI}{logical. Don't take that argument into account (parameter for the
 user interface) !}
}
\value{
A stack containing the environmental variables raster (normalized or
 not)
}
\description{
Function to load environmental variables data from rasters to do
\code{\link{Modelling}}, \code{\link{Ensemble.Modelling}} or
\code{\link{Stack.Modelling}}.
}
\examples{
\dontrun{
load.var(directory)
}
}
\seealso{
\code{\link{load.occ}} to load occurences
}

