% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GUI.R
\name{NumberOneGUI}
\alias{NumberOneGUI}
\title{Number one package Global User Interface}
\usage{
NumberOneGUI(browser = F, maxmem = 1e+10)
}
\arguments{
\item{browser}{logical. Option to plot or not the user global interface in
your internet browser}

\item{maxmem}{numeric. Option to choose the maximum memory allocated to the
  user interface}
}
\value{
Open a window with a shiny app allowing to use the number one package
  with an easy user interface
}
\description{
User interface to use the number one package
}
\details{
Due to the relatively important size of environmental data you
  should gave enough memory to the interface
}
\examples{
\dontrun{
NumberOneGUI()
}
}

