% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Occurences.R
\docType{data}
\name{Occurences}
\alias{Occurences}
\title{Plant occurences data frame}
\format{A data frame with 839 rows and 4 variables: \describe{
  \item{genus}{genus of the specie} \item{specie}{observed specie}
  \item{Longitude}{Longitude of the specie observation}
  \item{Latitude}{Latitude of the specie observation} }}
\usage{
Occurences
}
\description{
A dataset containing plant occurences from 7 species of New Caledonia.
Occurences location precision have been degraded to 5 kilometers. Occurences
data come from Noumea Herbarium (NOU) and NC-PIPN network (see Ibanez et al
in references below).
}
\references{
Ibanez, T., Munzinger, J., Dagostini, G., Hequet, V., Rigault,
  F., Jaffre, T. and Birnbaum, P. (2014) Structural and floristic
  characteristics of mixed rainforest in New Caledonia: new data from the New
  Caledonian Plant Inventory and Permanent Plot Network (NC-PIPPN). Applied
  Vegetation Science, 17, 386-397.
}
\keyword{datasets}

