% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Stack.Niche.Model.R
\docType{class}
\name{Stack.Species.Ensemble.Niche.Model-class}
\alias{Stack.Species.Ensemble.Niche.Model-class}
\title{An S4 class to represent a stack species distribution ensemble model of
multiple algorithms and multiple species}
\description{
This is an S4 class to represent a stack species distribution ensemble model
of multiple algorithms (among generalized linear model, general additive
model, multivariate adpatative splines, generalized boosted regression models,
classification tree analysis, random forest, maximum entropy, artificial
neural network, and support vector machines). It can be obtain with
\code{\link{Stack.Modelling}} or \code{\link{stacking}}.
}
\section{Slots}{

\describe{
\item{\code{name}}{character. Name of the model (by default
Stacked.Species.Niche.Model)}

\item{\code{diversity.map}}{raster. Local species richness map of the model}

\item{\code{uncertainity}}{raster. Intermodel variance map}

\item{\code{evaluation}}{data frame. Evaluation of the model (threshold, AUC, omission
rate, sensitivity, specificity, correct proportion and Kappa)}

\item{\code{algorithm.evluation}}{data frame. Evaluation of each algorihtm with the
same metrics as evaluation slot}

\item{\code{variables.importance}}{data frame. Relative percentage of importance for
each variable used in the model}

\item{\code{algorithm.correlation}}{data.frame. Correlation matrix inter algorithms
calculated with pearson coefficient on habitat suitability map}

\item{\code{enms}}{list. List with each ensemble model used for the stack model}

\item{\code{parameters}}{data frame. Parameters used to realized the model}
}}
\seealso{
\linkS4class{Ensemble.Niche.Model} an S4 class for ensemble models,
 and \linkS4class{Algorithm.Niche.Model} an S4 class for algorithm models.
}

