% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/update.stack.R
\docType{methods}
\name{update,Stack.Species.Ensemble.Niche.Model-method}
\alias{update,Stack.Species.Ensemble.Niche.Model-method}
\title{Update a previous stack}
\usage{
\S4method{update}{Stack.Species.Ensemble.Niche.Model}(object, Occurences, Env,
  Xcol = "Longitude", Ycol = "Latitude", Pcol = NULL, Spname = NULL,
  name = stack@name, save = F, directory = getwd(), thresh = 1001,
  tmp = F, verbose = T, GUI = F, ...)
}
\arguments{
\item{object}{Stack.Species.Ensemble.Niche.Model. The previously realized stack
species model}

\item{Occurences}{data frame. New or updated occurences table (can be treated
first by \code{\link{load.occ}}).}

\item{Env}{raster object. Environnment raster object (can be treated first by
\code{\link{load.var}}).}

\item{Xcol}{character. Name of the occurences table column containing Latitude
or X coordinates.}

\item{Ycol}{character. Name of the occurences table column containing
Longitude or Y coordinates.}

\item{Pcol}{character. Name of the occurences table column containing presence
or absence value, if NULL presence-only data set is assumed.}

\item{Spname}{character. Name of the new or updated specie}

\item{name}{character. Optionnal name given to the final
Stack.Specie.Ensemble.Niche.Model producted, by default it's the name of the
previous stack.}

\item{save}{logical. If true the model is automatically saved.}

\item{directory}{character. If save is true, the name of the directory to save
the model.}

\item{thresh}{numeric. binary map threshold computing precision parmeter, the
higher it is the more accurate is the threshold but the longer is the
modelling evaluation step !}

\item{tmp}{logical. If true ensemble models habitat suitability map and
uncertainty map rasters are saved in temporary files to release memory.}

\item{verbose}{logical. If true allow the function to print text in the
console}

\item{GUI}{logical. Don't take that argument into account (parameter for the
user interface) !}

\item{...}{additionnal parameters for the algorithm modelling function (see
 details below).}
}
\value{
an S4 \linkS4class{Stack.Species.Ensemble.Niche.Model} Class object
 viewable with \code{\link{plot.model}} method
}
\description{
Update a previous stack species ensemble distribution model with new
occurences datas. It takes in inputs an updated or new occurences data frame
from one specie, previous envrionmental variables, and an S4
\linkS4class{Stack.Species.Ensemble.Niche.Model} class object containing a
previously realized stack species model.
}
\examples{
\dontrun{
update(stack, Occurences, Env, Spname = 'NewSpecie')
}
}
\seealso{
\code{\link{Stack.Modelling}} for stack species ensemble modelling
 with multiple algorithms and species
}

