% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/load.model.R
\name{load.model}
\alias{load.enm}
\alias{load.model}
\alias{load.stack}
\title{Load models functions}
\usage{
load.enm(name, directory = getwd())

load.stack(name = "Stack", directory = getwd())
}
\arguments{
\item{name}{character. Folder name of the saved model.}

\item{directory}{character. Path to the directory containing the saved model folder, by default the current directory.}
}
\value{
The corresponding model object
}
\description{
Allow to load S4 Ensemble.Niche.Model and Stack.Species.Ensemble.Niche.Model objects saved with their respective save functions
}
\seealso{
save.model
}

