% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot.model.R
\docType{methods}
\name{plot.model}
\alias{plot.model}
\title{Plot models functions}
\usage{
\S4method{plot}{Stack.Species.Ensemble.Niche.Model,ANY}(x, y, ...)
}
\arguments{
\item{x}{Object to pe plot (S4 Algorithm.Niche.Model, Ensemble.Niche.Model or Stack.Species.Ensemble.Niche.Model object)}
}
\value{
Open a window with a shiny app rendering all the results in an easy way to see all results in one glance
}
\description{
Allow to plot S4 Algorithm.Niche.Model, Ensemble.Niche.Model and Stack.Species.Ensemble.Niche.Model objects.
}

