% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/save.model.R
\docType{methods}
\name{save.model}
\alias{save.enm,Ensemble.Niche.Model-method}
\alias{save.model}
\alias{save.stack,Stack.Species.Ensemble.Niche.Model-method}
\title{Save models functions}
\usage{
\S4method{save.enm}{Ensemble.Niche.Model}(enm, name = strsplit(enm@name, ".",
  fixed = T)[[1]][1], directory = getwd())

\S4method{save.stack}{Stack.Species.Ensemble.Niche.Model}(stack,
  name = "Stack", directory = getwd())
}
\arguments{
\item{enm}{Ensemble.Niche.Model. Ensemble model to be saved.}

\item{name}{character. Folder name of the saved model.}

\item{directory}{character. Path to the directory containing the saved model folder, by default the current directory.}

\item{stack}{Stack.Species.Ensemble.Niche.Model.Stack species ensemble models to be saved.}
}
\value{
Nothing in R environment. Create folders and tables and rasters associated to the model. Tables are in .csv and rasters in .grd/.gri.
}
\description{
Allow to save S4 Ensemble.Niche.Model and Stack.Species.Ensemble.Niche.Model objects.
}
\seealso{
load.model
}

